app_process: Integrate Xposed Framework modifications.
Obviously, credit goes to @rovo89 .
While I would have liked to keep a commit history, the way Xposed is
organized isn't ROM-integration-friendly for that to be feasible.
I don't think it's appropriate to set him as the commit author either
since he's clearly against integration, so the commit mention will
have to do.